Update OLS for date change
Changed the date range for the data sets to exclude the pandemic. We originally trained a model on prepandemic data and tested it on post pandemic data which was a bad idea considering the pandemic's unexpected, large effect it had on commodities. After changing the date range, our projections are more accurate.